ensemble: their n_sampled_cols is ONE variable; ours had become two
Found by reading builder.cuh and both split kernels against ours, not by a
failing check -- every check we had ran where the two agree.

builder.cuh:240 sets dataset.n_sampled_cols in the Builder ctor and :439-440
narrows it per sampling round, and that same field is BOTH the stride
sample_features writes at (:509 passes it as `k`) and the stride the split
kernels read at (kernels/builder_kernels_impl.cuh:310, :373). One variable,
so writer and reader cannot disagree.

Ours kept the sampled count on the Builder and left the DatasetView field at
whatever the caller passed -- randomforest.mojo hands it n_cols. Writer strode
by max(1, max_features*n_cols), reader by n_cols. Node 0 coincided; every
later node in a batch read a neighbour's drawn columns, and far enough into a
batch, past the end of column_samples. That is the DEFAULT classifier path:
cuML ships max_features='sqrt'.

The load-bearing write is the per-round one in do_split -- narrowing in train
alone is inert, because do_split overwrites it every round.

Two more places we had quietly improved on them:

  - n_sampled_rows: randomforest.cuh:302 is round(max_samples * n_rows) with
    max_samples a float and n_rows an int, so the product is FLOAT32. Ours
    computed it in Float64. Above 2^24 that is a different bootstrap size and
    so a different forest -- at n_rows 20,000,001 theirs gives 20,000,000.

  - score(): raft scores.cuh:116 is `double diff = predictions[i] -
    ref_predictions[i]` with both operands float, so the SUBTRACTION is
    float32. Ours widened first. The sums might be excused by atomic order;
    the median is a selection, so it can return a different element.

Also compute_quantiles: every scratch buffer's last use was the enqueue that
took its pointer, so Mojo's free-at-last-use left all of them dead under
kernels still queued to read them -- the defect already on record, which
passed a green run before it fired. Keep-alives after the synchronize.

sampled_cols_check is new and compares PER CELL against the sampler's own
output read back off the device, because a total cannot see placement. Its
sabotage arm restores the old stride and shows node 2 splitting on column 0
while its own draw was {12,10,5,8}.

All 17 checks pass.

+++ ensemble/randomforest.mojo
@@ -1032,7 +1032,15 @@ struct RandomForest[dtype: DType, label_dtype: DType](
         var abs_sum: Float64 = 0.0
         var sq_sum: Float64 = 0.0
         for i in range(n_rows):
-            var diff = Float64(predictions[i]) - Float64(ref_labels[i])
+            # raft `scores.cuh:116` -- `double diff = predictions[i] -
+            # ref_predictions[i];`. BOTH OPERANDS ARE `T`, which is `float`
+            # for `RandomForestRegressorF`, so THE SUBTRACTION IS FLOAT32
+            # and only its result is widened. Widening first is a silent
+            # improvement on them: the two sums might be excused by the
+            # atomic ordering already priced in DEVIATION 119e, but the
+            # median is a SELECTION, so a different low bit here returns a
+            # different element, not a differently-rounded one.
+            var diff = Float64(predictions[i] - ref_labels[i])
             var abs_diff = diff
             if abs_diff < 0.0:
                 abs_diff = -abs_diff
@@ -1471,8 +1479,17 @@ def n_sampled_rows_for(
     """
     if not bootstrap:
         return n_rows
-    var x = Float64(max_samples) * Float64(n_rows)
-    var f = Float64(Int(x))
+    # `:302` -- `std::round(this->rf_params.max_samples * n_rows)`.
+    # THE PRODUCT IS FLOAT32. `max_samples` is `float` (`randomforest.hpp:71`)
+    # and `n_rows` is `int`, so C++'s usual arithmetic conversions promote
+    # `n_rows` TO FLOAT and the multiply rounds there -- including the
+    # int->float conversion of `n_rows` itself. Computing this in Float64
+    # is not a harmless widening: above 2^24 float32 cannot hold every
+    # integer, so at n_rows = 20_000_001 with max_samples = 1.0 theirs
+    # gives 20_000_000 and a Float64 product gives 20_000_001. That is a
+    # different bootstrap size, hence a different forest, on every tree.
+    var x = Float32(max_samples) * Float32(n_rows)
+    var f = Float32(Int(x))
     if x - f >= 0.5:
         return Int(f) + 1
     return Int(f)
